@@ -1,23 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
         <LangVersion>latest</LangVersion>
         <TypeScriptToolsVersion>3.4.3</TypeScriptToolsVersion>
         <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
-        <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+        <NoWarn>1701;1702;1591</NoWarn>
+        <CodeAnalysisRuleSet>..\Vault.ruleset</CodeAnalysisRuleSet>
     </PropertyGroup>
 
     <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
         <DocumentationFile>bin\Debug\netcoreapp2.2\Vault.xml</DocumentationFile>
-        <CodeAnalysisRuleSet>$(SolutionDir)\Vault.ruleset</CodeAnalysisRuleSet>
-        <NoWarn>1701;1702;1591</NoWarn>
     </PropertyGroup>
 
     <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
         <DocumentationFile>bin\Release\netcoreapp2.2\Vault.xml</DocumentationFile>
-        <CodeAnalysisRuleSet>$(SolutionDir)\Vault.ruleset</CodeAnalysisRuleSet>
-        <NoWarn>1701;1702;1591</NoWarn>
     </PropertyGroup>
 
     <ItemGroup>
@@ -29,11 +26,12 @@
 
     <ItemGroup>
         <PackageReference Include="Dapper" Version="1.60.6" />
-        <PackageReference Include="Microsoft.AspNetCore.App" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.0" />
         <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="2.9.1">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
         </PackageReference>
+        <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
         <PackageReference Include="StyleCop.Analyzers" Version="1.1.0-beta008">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
